wip: extract agent subpackage and improve naming for step subpackage.
