+++ hub/step/step.go
@@ -1,4 +1,4 @@
-package steps
+package step
 
 import (
 	"fmt"
@@ -18,7 +18,7 @@ import (
 	"github.com/greenplum-db/gpupgrade/step"
 )
 
-func BeginStep(stateDir string, name string, sender idl.MessageSender) (*step.Step, error) {
+func Begin(stateDir string, name string, sender idl.MessageSender) (*step.Step, error) {
 	path := filepath.Join(stateDir, fmt.Sprintf("%s.log", name))
 	log, err := os.OpenFile(path, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0600)
 	if err != nil {
