+++ hub/step/step_test.go
@@ -1,4 +1,4 @@
-package steps
+package step
 
 import (
 	"bytes"
